Rename ses to aws-ses
