+++ alerting/alert/type.go
@@ -5,6 +5,9 @@ package alert
 type Type string
 
 const (
+	// TypeAWSSes is the Type for the aws-ses alerting provider
+	TypeAWSSes Type = "aws-ses"
+
 	// TypeCustom is the Type for the custom alerting provider
 	TypeCustom Type = "custom"
 
@@ -43,9 +46,6 @@ const (
 
 	// TypePushover is the Type for the pushover alerting provider
 	TypePushover Type = "pushover"
-
-	// TypeSes is the Type for the ses alerting provider
-	TypeSes Type = "ses"
 
 	// TypeSlack is the Type for the slack alerting provider
 	TypeSlack Type = "slack"
